+++ extensions/faker/FixtureController.php
@@ -101,20 +101,20 @@ use yii\helpers\VarDumper;
  * //list all templates under specified template path
  * yii fixture/templates --templatePath='@app/path/to/my/custom/templates'
  * ~~~
- * 
+ *
  * You also can create your own data providers for custom tables fields, see Faker library guide for more info (https://github.com/fzaninotto/Faker);
  * After you created custom provider, for example:
  *
  * ~~~
  * class Book extends \Faker\Provider\Base
  * {
- * 
+ *
  *     public function title($nbWords = 5)
  *     {
  *         $sentence = $this->generator->sentence($nbWords);
  *         return mb_substr($sentence, 0, mb_strlen($sentence) - 1);
  *     }
- * 
+ *
  * }
  * ~~~
  *
@@ -130,6 +130,8 @@ use yii\helpers\VarDumper;
  *        ],
  *    ],
  * ~~~
+ *
+ * @property \Faker\Generator $generator This property is read-only.
  *
  * @author Mark Jebri <mark.github@yandex.ru>
  * @since 2.0.0