+++ extensions/bootstrap/CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 bootstrap extension Change Log
 
 - Chg #5231: Collapse `items` property uses `label` element instead of array key for headers (nkovacs)
 - Chg #5232: Collapse encodes headers by default (nkovacs)
+- Chg #5217: Tabs no longer requires content since empty tab could be used dynamically (damiandennis)
 
 
 2.0.0-rc September 27, 2014